# swagger/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Exemplo WEBSERVICE",
    "version" : "v0.1",
    "title" : "Exemplo webservice",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/webservice",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/getPessoa/{nomePessoa}" : {
      "get" : {
        "operationId" : "getPessoa",
        "parameters" : [ {
          "name" : "nomePessoa",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PessoaModel"
            }
          }
        }
      }
    },
    "/getPessoas" : {
      "get" : {
        "operationId" : "getPessoas",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PessoaModel"
              }
            }
          }
        }
      }
    },
    "/getPessoasPorAnoNascimento/{ano}" : {
      "get" : {
        "operationId" : "getPessoaPorAnoNascimento",
        "parameters" : [ {
          "name" : "ano",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PessoaModel"
              }
            }
          }
        }
      }
    },
    "/getPessoasPorNaturalidade/{naturalidadePessoa}" : {
      "get" : {
        "operationId" : "getPessoaPorNaturalidade",
        "parameters" : [ {
          "name" : "naturalidadePessoa",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PessoaModel"
              }
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "PessoaModel" : {
      "type" : "object",
      "properties" : {
        "nome" : {
          "type" : "string"
        },
        "naturalidade" : {
          "type" : "string"
        },
        "anoNascimento" : {
          "type" : "string"
        }
      }
    }
  }
}